requirements.txt
django
python-dotenv
langchain
langchain-community
faiss-cpu
PyMuPDF
google-generativeai
